+++ src/main/java/org/abos/twi/gatcha/core/battle/Effect.java
@@ -1,11 +1,11 @@
 package org.abos.twi.gatcha.core.battle;
 
-import org.abos.common.Vec2d;
+import org.abos.common.Vec2i;
 
 public interface Effect {
 
     EffectType getEffectType();
 
-    void apply(final CharacterInBattle from, final Vec2d target, final Field field);
+    void apply(final CharacterInBattle from, final Vec2i target, final Field field);
 
 }
